Correct deleteTrickFiles which tried to delete public directory if no main image is present

+++ src/Controller/TrickController.php
@@ -419,7 +419,7 @@ class TrickController extends AbstractController
     {
         $rootDirectory = dirname(dirname(__DIR__));
 
-        if (strpos($trick->getMainImage(), "http") === false) {
+        if ($trick->getMainImage() && strpos($trick->getMainImage(), "http") === false) {
             unlink($rootDirectory . "/public" . $trick->getMainImage());
         }
